requirements.txt
PyQt6
torch
torchvision
basicsr
realesrgan
facexlib
gfpgan
numpy
opencv-python-headless
Pillow
